backend/requirements.txt
fastapi
uvicorn
boto3
pydantic
pydantic-settings
python-dotenv
sqlmodel
aiosqlite
httpx
pytest
google-genai
openai